requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
py-cord==2.4.1
pycparser==2.21
PyNaCl==1.5.0
sniffio==1.3.0
typing_extensions==4.7.1
voicevox-client==0.2.0
yarl==1.9.2
